Adds .log
