+++ .gitignore
@@ -14,3 +14,4 @@ docs/build
 notebooks
 .mypy_cache
 .devcontainer/*
+
